Move mobile files under `mobile` dir
