+++ auth0_flutter/lib/src/mobile/authentication_api.dart
