+++ auth0_flutter/lib/src/mobile/credentials_manager.dart
